SystemManager, Transform Hierarchy, Quaternions, Application Clean-up

+++ ImaGE-Editor/source/GUI/Dockable/SceneHierarchy.cpp
@@ -8,6 +8,7 @@
 #include <ImGui/misc/cpp/imgui_stdlib.h>
 #include <Prefabs/PrefabManager.h>
 #include <GUI/Dockable/Inspector.h>
+#include <Core/Systems/TransformSystem/TransformHelpers.h>
 
 namespace GUI
 {
@@ -16,7 +17,7 @@ namespace GUI
     mEntityManager{ ECS::EntityManager::GetInstance() },
     mSceneName{}, 
     mRightClickedEntity{}, mRightClickMenu{ false }, mEntityOptionsMenu{ false },
-    mPrefabPopup{ false }, mFirstTimePfbPopup{ true }, mEditingPrefab{ false }, lockControls{ false }
+    mPrefabPopup{ false }, mFirstTimePfbPopup{ true }, mEditingPrefab{ false }, mLockControls{ false }, mSceneModified{ false }
   {
     SUBSCRIBE_CLASS_FUNC(Events::EventType::SCENE_STATE_CHANGE, &SceneHierarchy::HandleEvent, this);
     SUBSCRIBE_CLASS_FUNC(Events::EventType::EDIT_PREFAB, &SceneHierarchy::HandleEvent, this);
@@ -58,8 +59,17 @@ namespace GUI
     if (ImGuiHelpers::BeginDrapDropTargetWindow(sDragDropPayload))
     {
       ImGuiPayload const* drop{ ImGui::AcceptDragDropPayload(sDragDropPayload) };
-      ECS::Entity const droppedEntity{ *reinterpret_cast<ECS::Entity*>(drop->Data) };
-      mEntityManager.RemoveParent(droppedEntity);
+      ECS::Entity droppedEntity{ *reinterpret_cast<ECS::Entity*>(drop->Data) };
+      if (mEntityManager.HasParent(droppedEntity)) {
+        mEntityManager.RemoveParent(droppedEntity);
+
+        // entity unparented - convert to local and recompute matrix
+        Component::Transform& trans{ droppedEntity.GetComponent<Component::Transform>() };
+        trans.SetLocalToWorld();
+        trans.modified = true;
+
+        QUEUE_EVENT(Events::SceneModifiedEvent);
+      }
 
       ImGui::EndDragDropTarget();
     }
@@ -71,7 +81,7 @@ namespace GUI
       }
     }
 
-    if (ImGui::IsKeyPressed(ImGuiKey_Delete) && GUIManager::GetSelectedEntity() && !lockControls) {
+    if (ImGui::IsKeyPressed(ImGuiKey_Delete) && GUIManager::GetSelectedEntity() && !mLockControls) {
       ECS::EntityManager::GetInstance().RemoveEntity(GUIManager::GetSelectedEntity());
       GUIManager::SetSelectedEntity(ECS::Entity());
     }
@@ -112,7 +122,7 @@ namespace GUI
       case Events::SceneStateChange::NEW:
       case Events::SceneStateChange::CHANGED:
         mSceneName = sceneStateEvent->mSceneName;
-        mEditingPrefab = false;
+        mEditingPrefab = mSceneModified = false;
         break;
       case Events::SceneStateChange::STOPPED:
         mEditingPrefab = false;
@@ -124,7 +134,10 @@ namespace GUI
       break;
     }
     case Events::EventType::SCENE_MODIFIED:
+      if (mSceneModified) { return; }
+
       mSceneName += " *";
+      mSceneModified = true;
       break;
     case Events::EventType::SAVE_SCENE:
       mSceneName.erase(mSceneName.size() - 2);
@@ -168,7 +181,7 @@ namespace GUI
         }
         if (ImGui::InputText("##EntityRename", &entityName, ImGuiInputTextFlags_EnterReturnsTrue | ImGuiInputTextFlags_AutoSelectAll)) {
           entity.GetComponent<Component::Tag>().tag = entityName;
-          editNameMode = lockControls = false;
+          editNameMode = mLockControls = false;
           firstEnterEditMode = true;
           QUEUE_EVENT(Events::SceneModifiedEvent);
         }
@@ -176,13 +189,13 @@ namespace GUI
 
         if (ImGui::IsKeyPressed(ImGuiKey_Escape)) {
           ImGui::SetWindowFocus(NULL);
-          editNameMode = lockControls = false;
+          editNameMode = mLockControls = false;
           firstEnterEditMode = true;
         }
       }
 
       bool dragNDropped{ false };
-      if (!lockControls) {
+      if (!mLockControls) {
         if (ImGui::BeginDragDropSource())
         {
           ECS::EntityManager::EntityID id{ entity.GetRawEnttEntityID() };
@@ -197,11 +210,15 @@ namespace GUI
           ImGuiPayload const* drop = ImGui::AcceptDragDropPayload(sDragDropPayload);
           if (drop)
           {
-            ECS::Entity const droppedEntity{ *reinterpret_cast<ECS::EntityManager::EntityID*>(drop->Data) };
+            ECS::Entity droppedEntity{ *reinterpret_cast<ECS::EntityManager::EntityID*>(drop->Data) };
             if (mEntityManager.HasParent(droppedEntity)) {
               mEntityManager.RemoveParent(droppedEntity);
             }
             mEntityManager.SetParentEntity(entity, droppedEntity);
+            droppedEntity.GetComponent<Component::Transform>().modified = true;
+            // entity has new parent, traverse down hierarchy and update transforms
+            TransformHelpers::UpdateTransformToNewParent(droppedEntity);
+            QUEUE_EVENT(Events::SceneModifiedEvent);
           }
           ImGui::EndDragDropTarget();
         }
@@ -215,19 +232,17 @@ namespace GUI
       if (ImGui::IsMouseReleased(ImGuiMouseButton_Left) && ImGui::IsItemHovered()) {
         if (isCurrentEntity) {
           if (!dragNDropped) {
-            editNameMode = lockControls = true;
+            editNameMode = mLockControls = true;
           }
         }
         else {
           GUIManager::SetSelectedEntity(entity);
-          editNameMode = lockControls = false;
+          editNameMode = mLockControls = false;
         }
       }
 
-      if (hasChildren)
-      {
-        for (auto const& child : mEntityManager.GetChildEntity(entity))
-        {
+      if (hasChildren) {
+        for (auto const& child : mEntityManager.GetChildEntity(entity)) {
           RecurseDownHierarchy(child);
         }
       }
